notify_bot/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==3.6.2
    # via
    #   -r requirements.in
    #   httpcore
cachetools==5.3.0
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.1.0
    # via
    #   -r requirements.in
    #   requests
h11==0.14.0
    # via
    #   -r requirements.in
    #   httpcore
h2==4.1.0
    # via -r requirements.in
hpack==4.0.0
    # via
    #   -r requirements.in
    #   h2
httpcore==0.16.3
    # via
    #   -r requirements.in
    #   httpx
httpx==0.23.3
    # via
    #   -r requirements.in
    #   python-telegram-bot
hyperframe==6.0.1
    # via
    #   -r requirements.in
    #   h2
idna==3.4
    # via
    #   -r requirements.in
    #   anyio
    #   requests
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
python-telegram-bot==20.2
    # via -r requirements.in
requests==2.28.2
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via
    #   -r requirements.in
    #   httpx
sniffio==1.3.0
    # via
    #   -r requirements.in
    #   anyio
    #   httpcore
    #   httpx
urllib3==1.26.15
    # via
    #   -r requirements.in
    #   requests
